feat(update): prompt for cert if auto update requires one
